合并更改

+++ interface/SGP40.py
@@ -86,13 +86,13 @@ class SGP40:
     def measureRaw(self, temperature, humidity):
         # 2*humi + CRC
         # paramh = struct.pack(">H", math.ceil(humidity * 0xffff / 100))
-        h = humidity * 0xFFFF / 100
+        h = int(humidity * 0xFFFF / 100)
         paramh = (h >> 8, h & 0xFF)
         crch = self.__crc(paramh[0], paramh[1])
 
         # 2*temp + CRC
         # paramt = struct.pack(">H", math.ceil((temperature + 45) * 0xffff / 175))
-        t = (temperature + 45) * 0xFFFF / 175
+        t = int((temperature + 45) * 0xFFFF / 175)
         paramt = (t >> 8, t & 0xFF)
         crct = self.__crc(paramt[0], paramt[1])
